Chart response of file paths
