+++ probot/index.js
@@ -4,7 +4,7 @@
  */
 
 import {sendRepo} from './components/sendRepo.js';
-import {sendRatingsAndComment} from './components/sendRatings.js';
+import {sendRatings} from './components/sendRatings.js';
 import axios from "axios";
 
 
@@ -59,69 +59,143 @@ export default (app) => {
 			console.error('Failed to process repositories:', error);
 		}
 	});
-// Handler for issues.opened event
-	app.on('issues.opened', async (context) => {
-		const issue = context.payload.issue;
-		const repository = context.payload.repository;
 
-		console.log(`Issue #${issue.number} opened in repository ${repository.full_name}`);
 
-		const issueAuthor = issue.user;
-		if (issueAuthor.type === 'Bot') {
-			console.log(`Issue #${issue.number} opened by a bot (${issueAuthor.login}). Ignoring.`);
-			return; // Exit early if the issue was opened by a bot
-		}
+	// Handler for issues.opened event
+app.on('issues.opened', async (context) => {
+    const issue = context.payload.issue;
+    const repository = context.payload.repository;
 
-		// Prepare data to send to Flask backend
-		const issueBody = issue.body || issue.title;
+    console.log(`Issue #${issue.number} opened in repository ${repository.full_name}`);
 
-		try {
-			// Fetch the full repository data to ensure all necessary fields are present
-			const {data: fullRepo} = await context.octokit.repos.get({
-				owner: repository.owner.login,
-				repo: repository.name,
-			});
+    const issueAuthor = issue.user;
+    if (issueAuthor.type === 'Bot') {
+        console.log(`Issue #${issue.number} opened by a bot (${issueAuthor.login}). Ignoring.`);
+        return; // Exit early if the issue was opened by a bot
+    }
 
-			// Pass the full repository object and context to sendRepo
-			const repoData = await sendRepo(fullRepo, context);
-			const fullData = {
-				issue: issueBody,
-				repository: repoData,
-			};
+    // Prepare data to send to Flask backend
+    const issueBody = issue.body || issue.title;
 
-			if (!fullData) {
-				console.error(`sendRepo returned null for ${repository.full_name}. Skipping Axios POST.`);
-				// Optionally, you can create a fallback comment here
-				const fallbackComment = 'Thank you for opening this issue! However, we encountered an error while processing it. Please try again later.';
-				await sendRatingsAndComment(null, context, fallbackComment, issueBody);
-			} else {
-				try {
-					const flaskResponse = await axios.post('http://localhost:5000/report', fullData, {
-						headers: {
-							'Content-Type': 'application/json',
-						},
-					});
+    try {
+        // Fetch the full repository data to ensure all necessary fields are present
+        const { data: fullRepo } = await context.octokit.repos.get({
+            owner: repository.owner.login,
+            repo: repository.name,
+        });
 
-					if (flaskResponse.status !== 200) {
-						throw new Error(`Failed to send data to Flask backend: ${flaskResponse.status} ${flaskResponse.statusText}`);
-					}
+        // Pass the full repository object and context to sendRepo
+        const repoData = await sendRepo(fullRepo, context);
+        const fullData = {
+            issue: issueBody,
+            repository: repoData,
+        };
 
-					console.log('Repo info sent to Flask backend successfully from issues.opened event.');
+        if (!fullData) {
+            console.error(`sendRepo returned null for ${repository.full_name}. Skipping Axios POST.`);
+        } else {
+            try {
+                const flaskResponse = await axios.post('http://localhost:5000/report', fullData, {
+                    headers: {
+                        'Content-Type': 'application/json',
+                    },
+                });
 
-					// Call the consolidated sendRatingsAndComment function
-					await sendRatingsAndComment(flaskResponse.data, context, 'Thank you for opening this issue!', issueBody);
-				} catch (error) {
-					console.error('Error while sending repo info from issues.opened:', error);
-					// Even if the Flask call fails, send a fallback comment
-					const errorComment = 'Thank you for opening this issue! We encountered an error while processing your request. Please try again later.';
-					await sendRatingsAndComment(null, context, errorComment, issueBody);
-				}
-			}
-		} catch (error) {
-			console.error(`Failed to fetch repository data for ${repository.full_name}:`, error);
-			// Create a comment indicating a failure to process the issue
-			const fallbackComment = 'Thank you for opening this issue! However, we encountered an error while processing it. Please try again later.';
-			await sendRatingsAndComment(null, context, fallbackComment, issueBody);
-		}
-	});
+                if (flaskResponse.status !== 200) {
+                    throw new Error(`Failed to send data to Flask backend: ${flaskResponse.status} ${flaskResponse.statusText}`);
+                }
+
+                console.log('Repo info sent to Flask backend successfully from issues.opened event.', flaskResponse.data['ranked_files']);
+                sendRatings(flaskResponse.data, context);
+
+                // Extract preprocessed bug report from the Flask response
+                const preprocessedBugReport = flaskResponse.data.preprocessed_bug_report;
+
+                // Now, construct the issueCommentBody including the preprocessed bug report
+                let issueCommentBody = 'Thank you for opening this issue!';
+
+                if (preprocessedBugReport) {
+                    issueCommentBody += `\n\n**Preprocessed Bug Report:**\n\n${preprocessedBugReport}`;
+                } else {
+                    issueCommentBody += '\n\nNo preprocessed bug report was returned.';
+                }
+
+                // Check for images in the issue body
+                if (issueBody) {
+                    const imageRegex = /https?:\/\/[^\s)\]]+/g;
+                    const imageUrls = issueBody.match(imageRegex);
+
+                    if (imageUrls && imageUrls.length > 0) {
+                        issueCommentBody += '\n\nI found the following images in the issue description:';
+                        imageUrls.forEach((imageUrl) => {
+                            issueCommentBody += `\n\n![Image](${imageUrl})`;
+                        });
+                    } else {
+                        issueCommentBody += '\n\nI did not find any images in the issue description.';
+                    }
+                }
+
+                // Handle creating or updating the comment
+                await handleIssueComment(context, issueCommentBody);
+            } catch (error) {
+                console.error('Error while sending repo info from issues.opened:', error);
+
+                // Even if the Flask call fails, we might want to still acknowledge the issue
+                let issueCommentBody = 'Thank you for opening this issue!';
+
+                // Check for images in the issue body
+                if (issueBody) {
+                    const imageRegex = /https?:\/\/[^\s)\]]+/g;
+                    const imageUrls = issueBody.match(imageRegex);
+
+                    if (imageUrls && imageUrls.length > 0) {
+                        issueCommentBody += '\n\nI found the following images in the issue description:';
+                        imageUrls.forEach((imageUrl) => {
+                            issueCommentBody += `\n\n![Image](${imageUrl})`;
+                        });
+                    } else {
+                        issueCommentBody += '\n\nI did not find any images in the issue description.';
+                    }
+                }
+
+                // Handle creating or updating the comment
+                await handleIssueComment(context, issueCommentBody);
+            }
+        }
+    } catch (error) {
+        console.error(`Failed to fetch repository data for ${repository.full_name}:`, error);
+    }
+});
+
+// Function to handle creating or updating the issue comment
+async function handleIssueComment(context, issueCommentBody) {
+    try {
+        // Check if the bot already commented
+        const comments = await context.octokit.issues.listComments(context.issue());
+        const botComment = comments.data.find(
+            (comment) => comment.user.type === 'Bot' && comment.body.includes('Thank you for opening this issue!')
+        );
+
+        if (botComment) {
+            const updatedCommentBody = 'EDIT: The bug report was updated.\n\n' + issueCommentBody;
+
+            console.log('Editing comment with ID:', botComment.id);
+            await context.octokit.issues.updateComment({
+                ...context.issue(),
+                comment_id: botComment.id,
+                body: updatedCommentBody,
+            });
+        } else {
+            // No bot comment exists, create a new one
+            const issueComment = context.issue({ body: issueCommentBody });
+
+            console.log('Creating comment:', issueCommentBody);
+            await context.octokit.issues.createComment(issueComment);
+        }
+    } catch (error) {
+        console.error('Error handling issue comments:', error);
+    }
+}
+
+
 };
